pyproject.toml
[project]
name = "langgraph-tool-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "chromadb>=1.0.15",
    "flask>=3.1.1",
    "ipykernel>=6.30.0",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.0",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "wikipedia>=1.4.0",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["*"]
exclude = ["data*", "vector_db*", "templates*", "__pycache__*", ".langgraph_api*", ".venv*", ".git*"]
